Migration fix 3 (#2)
* Migration fix

* Migration fix 2

* Migration fix 3

+++ migrations/m160620_131811_vote.php
@@ -13,7 +13,7 @@ class m160620_131811_vote extends Migration
             'id' => $this->primaryKey(),
             'entity' => $this->integer()->unsigned()->notNull(),
             'target_id' => $this->integer()->notNull(),
-            'user_id' => $this->integer(),
+            'user_id' => $this->bigInteger(),
             'user_ip' => $this->string(39)->notNull()->defaultValue('127.0.0.1'),
             'value' => $this->boolean()->defaultValue(false),
             'created_at' => $this->integer()->notNull(),
